Extracts LegacyMockInterface
